backend/services/image_processor.py
import logging
import base64
import io
from PIL import Image
import os
import tempfile

logger = logging.getLogger(__name__)

def extract_text_from_image(base64_image):
    """
    Extract text from a base64-encoded image
    
    In a production app, this would use OCR like Tesseract or a cloud API,
    but for demo purposes we'll return placeholder text
    """
    logger.info("Extracting text from image")
    
    try:
        # Decode base64 string to image
        image_data = base64.b64decode(base64_image)
        image = Image.open(io.BytesIO(image_data))
        
        # In a real implementation, we would process the image with OCR
        # For example, with pytesseract:
        # import pytesseract
        # extracted_text = pytesseract.image_to_string(image)
        
        # For demo purposes, return placeholder
        width, height = image.size
        extracted_text = f"This is placeholder text that would be extracted from your {width}x{height} image using OCR technology.\n\nIn a production app, this would contain the actual text from your image."
        
        logger.info("Image text extraction complete")
        return extracted_text
        
    except Exception as e:
        logger.error(f"Error processing image: {str(e)}")
        raise