backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
python-dotenv==1.0.0
transformers==4.36.2
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.0
SpeechRecognition==3.10.0
requests==2.31.0
pydub==0.25.1